Split code into store/types/wallet
